backend/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
mysqlclient>=2.2
celery>=5.3
redis>=5.0
django-cors-headers>=4.3
django-filter>=24.0
python-dotenv>=1.0
